requirements.txt
gradio
groq
python_dotenv
gTTS